Fix type

+++ Docker_Preparation_For_Submission/inference.py
@@ -460,7 +460,7 @@ def run():
         run_batch()
     metric._calculate_metrics()
 
-    mkotyushev_prob_masks = metric.preds.cpu().numpy()
+    mkotyushev_prob_masks = metric.preds.float().cpu().numpy()
 
     ###########################################################################
